# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for CineMatch movie platform",
        "title": "CineMatch API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@cinematch.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/streaming-services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated list of streaming services from the master table",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming-services"
                ],
                "summary": "Get all streaming services",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cursor for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 20,
                        "description": "Number of items",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.StreamingServicesListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replaces user's streaming services with the provided list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming-services"
                ],
                "summary": "Update user's streaming services (bulk)",
                "parameters": [
                    {
                        "description": "List of streaming service IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/streaming_services.UpdateStreamingServicesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/streaming-services/bulk": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes multiple streaming services from user's selected services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming-services"
                ],
                "summary": "Remove multiple streaming services from user's profile",
                "parameters": [
                    {
                        "description": "List of streaming service IDs to remove",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/streaming_services.UpdateStreamingServicesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/streaming-services/mine": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the streaming services selected by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming-services"
                ],
                "summary": "Get user's selected streaming services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.UserStreamingServicesListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/streaming-services/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Searches streaming services by name with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming-services"
                ],
                "summary": "Search streaming services",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cursor for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 20,
                        "description": "Number of items",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.SearchStreamingServicesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/streaming-services/{serviceId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a single streaming service from user's selected services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaming-services"
                ],
                "summary": "Remove a streaming service from user's profile",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Streaming service ID to remove",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/streaming_services.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "streaming_services.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "streaming_services.SearchStreamingServicesResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next_cursor": {
                    "type": "string"
                },
                "streamingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/streaming_services.StreamingServiceResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "streaming_services.StreamingServiceResponse": {
            "type": "object",
            "properties": {
                "iconUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "streaming_services.StreamingServicesListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "next_cursor": {
                    "type": "string"
                },
                "streamingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/streaming_services.StreamingServiceResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "streaming_services.UpdateStreamingServicesRequest": {
            "type": "object",
            "required": [
                "serviceIds"
            ],
            "properties": {
                "serviceIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "streaming_services.UserStreamingServiceResponse": {
            "type": "object",
            "properties": {
                "iconUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sourceId": {
                    "description": "Foreign key to streaming_services table",
                    "type": "integer"
                }
            }
        },
        "streaming_services.UserStreamingServicesListResponse": {
            "type": "object",
            "properties": {
                "streamingServices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/streaming_services.UserStreamingServiceResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by your access token (e.g., \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}